Implementation of task 3.2

+++ database.py
@@ -6,11 +6,28 @@ import os
 from btree import Btree
 import shutil
 from misc import split_condition
+from users import User
+from groups import Group
 
 class Database:
     '''
     Database class contains tables.
     '''
+        #Added
+        self.users = {}
+        self.groups = {}
+        self.loggedInAs = None
+        all_privs = {'Create_User' : True, 'Delete_Tbl' : True}
+        #admin
+        admin = User('admin', 'admin', all_privs, set(), set())
+        self.users.update({'admin': admin})
+        self.loggedInAs = admin
+        #groups
+        admin_group = Group('admins', {'Create_User' : True, 'Delete_Tbl' : True}, set(), set())
+        self.groups.update({'admins': admin_group})
+        viewer_group = Group('admins', {}, set(), set())
+        self.groups.update({'viewers': viewer_group})
+        #Added
 
     def __init__(self, name, load=True):
         self.tables = {}
@@ -42,6 +59,106 @@ class Database:
         self.create_table('meta_insert_stack',  ['table_name', 'indexes'], [str, list])
         self.create_table('meta_indexes',  ['table_name', 'index_name'], [str, str])
         self.save()
+
+        def login(self, username, password):
+        if self.loggedInAs == None:
+            usr = self.users.get(username)
+            if usr != None:
+                if usr.password == password:
+                    self.loggedInAs = usr
+                    print("Welcome ", username)
+                else:
+                    print('Username and password do not match')
+            else:
+                print('This user not found')
+        else:
+            print("Already Logged in.")
+
+    def logout(self):
+        self.loggedInAs = None
+        print("Logged Out Successfully")
+
+    def create_group(self, name, privs, del_privs, sel_privs):
+        if self.loggedInAs.can_create():
+            del_privileges = dict()
+            sel_privileges = dict()
+
+            for priv in del_privs:
+                del_privileges.update({priv: True})
+
+            for priv in sel_privs:
+                sel_privileges.update({priv: True})
+            self.groups.update({name: Group(name, privs, del_privileges, sel_privileges)})
+            print("{g} Group has been created".format(g=name))
+        else:
+            print("You don't have the privileges to create a new Group")
+
+    def add_user_to_group(self, group_name, user_name):
+        if self.loggedInAs.can_create():
+            group = self.groups.get(group_name)
+            user = self.users.get(user_name)
+            if group != None and user != None:
+                group.add_user(user)
+
+
+    def create_user(self,username, password):
+        priv = {'Create_User' : False, 'Delete_Tbl' : False}
+        if self.loggedInAs.can_create():
+            self.users.update({username: User(username, password, priv, set(), set())})
+            print("User: {U} has been created".format(U=username))
+        else:
+            print("You don't have the privileges to Create a new User")
+
+    def grant_all_privileges(self, username):
+        if self.loggedInAs.can_create():
+            user = self.users.get(username)
+            if user != None:
+                priv_up = {'Create_User' : True, 'Delete_Tbl' : True}
+                user.privileges.update(priv_up)
+
+    def grant_create_privilege(self, usr):
+        if self.loggedInAs.can_create():
+            user = self.users.get(usr)
+            if user != None:
+                priv_up = {'Create_User' : True}
+                user.privileges.update(priv_up)
+        else:
+            print("You dont have privileges to add privileges")
+
+    def grant_delete_privilege(self, usr):
+        if self.loggedInAs.can_create():
+            user = self.users.get(usr)
+            if user != None:
+                priv_up = {'Delete_Tbl' : True}
+                user.privileges.update(priv_up)
+                user.tbls_del.add('meta_locks')
+                user.tbls_del.add('meta_length')
+                user.tbls_del.add('meta_insert_stack')
+                print("Granted ", usr, " delete priv and required priv")
+        else:
+            print("You dont have privileges to add privileges")
+
+    def grant_select_table_privileges(self, usr, table):
+        if self.loggedInAs.can_create():
+            user = self.users.get(usr)
+            if user != None:
+                user.tbls_sel.add(table)
+                print("Granted ", usr, " required privileges to select table")
+                if 'meta_locks' not in user.tbls_sel:
+                    user.tbls_sel.add('meta_locks')
+        else:
+            print("You dont have privileges to add privileges")
+
+    def grant_delete_table_privileges(self, usr, table):
+        if self.loggedInAs.can_create():
+            user = self.users.get(usr)
+            if user != None:
+                user.tbls_del.add(table)
+                user.tbls_sel.add('meta_locks')
+                print("Granted ", usr, " required privileges to delete table rows")
+        else:
+            print("You dont have privileges to add privileges")
+
 
 
 
@@ -96,6 +213,16 @@ class Database:
         or
         db_object.table_name
         '''
+
+         # adding all privs for new table to admin
+        self.users.get('admin').tbls_sel.add(name)
+        self.users.get('admin').tbls_del.add(name)
+        # adding privileges to the appropriate groups
+        self.groups.get('admins').update_delete_privs(name)
+        self.groups.get('admins').update_select_privs(name)
+        self.groups.get('viewers').update_select_privs(name)
+
+
         self.tables.update({name: Table(name=name, column_names=column_names, column_types=column_types, primary_key=primary_key, load=load)})
         # self._name = Table(name=name, column_names=column_names, column_types=column_types, load=load)
         # check that new dynamic var doesnt exist already
@@ -113,23 +240,25 @@ class Database:
         '''
         Drop table with name 'table_name' from current db
         '''
-        self.load(self.savedir)
-        if self.is_locked(table_name):
-            return
+        if self.loggedInAs.can_delete():
+            self.load(self.savedir)
+            if self.is_locked(table_name):
+                return
 
-        self.tables.pop(table_name)
-        delattr(self, table_name)
-        if os.path.isfile(f'{self.savedir}/{table_name}.pkl'):
-            os.remove(f'{self.savedir}/{table_name}.pkl')
+      self.tables.pop(table_name)
+            delattr(self, table_name)
+            if os.path.isfile(f'{self.savedir}/{table_name}.pkl'):
+                os.remove(f'{self.savedir}/{table_name}.pkl')
+            else:
+                print(f'"{self.savedir}/{table_name}.pkl" does not exist.')
+            self.delete('meta_locks', f'table_name=={table_name}')
+            self.delete('meta_length', f'table_name=={table_name}')
+            self.delete('meta_insert_stack', f'table_name=={table_name}')
+
+       # self._update()
+            self.save()
         else:
-            print(f'"{self.savedir}/{table_name}.pkl" does not exist.')
-        self.delete('meta_locks', f'table_name=={table_name}')
-        self.delete('meta_length', f'table_name=={table_name}')
-        self.delete('meta_insert_stack', f'table_name=={table_name}')
-
-        # self._update()
-        self.save()
-
+            print('You do not have acess to drop table')
 
     def table_from_csv(self, filename, name=None, column_types=None, primary_key=None):
         '''
@@ -276,18 +405,21 @@ class Database:
 
                     operatores supported -> (<,<=,==,>=,>)
         '''
-        self.load(self.savedir)
-        if self.is_locked(table_name):
-            return
-        self.lockX_table(table_name)
-        deleted = self.tables[table_name]._delete_where(condition)
-        self.unlock_table(table_name)
-        self._update()
-        self.save()
-        # we need the save above to avoid loading the old database that still contains the deleted elements
-        if table_name[:4]!='meta':
-            self._add_to_insert_stack(table_name, deleted)
-        self.save()
+        if self.loggedInAs.can_del_tbl(table_name):
+            self.load(self.savedir)
+            if self.is_locked(table_name):
+                return
+            self.lockX_table(table_name)
+            deleted = self.tables[table_name]._delete_where(condition)
+            self.unlock_table(table_name)
+            self._update()
+            self.save()
+            # we need the save above to avoid loading the old database that still contains the deleted elements
+            if table_name[:4]!='meta':
+                self._add_to_insert_stack(table_name, deleted)
+            self.save()
+        else:
+            print("Cant delete rows from ", table_name)
 
     def select(self, table_name, columns, condition=None, order_by=None, asc=False,\
                top_k=None, save_as=None, return_object=False):
